project_backup_20260401_1100/requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.9

# Data Processing
pyyaml>=6.0.1
requests>=2.31.0
jinja2>=3.1.2

# AI/ML (CPU optimized for compatibility)
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Audio Processing
librosa>=0.10.1
soundfile>=0.12.1

# Text-to-Speech (Python 3.13 compatible)
pyttsx3>=2.90

# Computer Vision
opencv-python>=4.9.0.80

# NumPy for audio/ML
numpy>=2.0.0
